man2/, man3/, man-pages.7: Move VERSIONS next to STANDARDS
VERSIONS and STANDARDS are closely related (and often the distinction is
not so clear).  Now that we're going to add another section, HISTORY,
that is related to both, it makes sense to have the three together.

As a curiosity, the list in man-pages(7) that detailed what each section
should contain had them by accident(?) in the order that we're moving
to, instead of the order that was used elsewhere.

Signed-off-by: Alejandro Colomar <alx@kernel.org>