*.mk, CONTRIBUTING, INSTALL: lint, build, check: Reorganize some targets
Some targets which were under lint-* were really building cat pages, so
let's call it build-catman, since it's what it is.  As part of the
build, it will report warnings, of course, as any other build system, so
nothing really changed, except for the target names, and the path in the
build tree where the cat pages (and intermediate files) are placed,
which is now directly under <.tmp/man/*>.

Some other targets were checking that the cat pages were correct after
the build, so those targets have been moved to check-* targets.

Document that contributors should run both the 'lint' and 'check'
targets to check the correctness of their patches.

`make all`, a.k.a. `make build`, now builds _all_ that can be built,
including cat pages, and C programs.

Implementation detail: $LINTMAN has been renamed, since now it's used
also for things that are not linters.  Call it $NONSO_MAN, since it's a
list of the non-'.so' man pages, which are the ones we want to lint,
build, and check.

Future directions:

I plan to implement 'build-html' using groff(1), which will reuse part
of the build-catman pipeline.  That will produce much higher quality
HTML manual pages.

Cc: G. Branden Robinson <g.branden.robinson@gmail.com>
Cc: Elliott Hughes <enh@google.com>
Cc: Oskari Pirhonen <xxc3ncoredxx@gmail.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>