verbose.mk: Fix warning from GNU Make 4.4
While '$(V).SILENT:' is a common idiom in make, it may be more explicit
to put '.SILENT:' inside a conditional.  Since we already used the
conditional for something else, it's not a big change.  As a nice side
effect, vim now recognizes it and highlights it as a special target.

With the old code, GNU Make 4.4 reported a warning about undefined
variables:

	lib/verbose.mk:18: warning: undefined variable 'V'

Signed-off-by: Alejandro Colomar <alx@kernel.org>