bash_aliases: Optimize man_section()
mandoc(1) renders pages much faster than groff(1), which is itself much
faster than using man(1).  This might seem irrelevant for a single page,
but this function is called in a loop in man_lsfunc() and man_lsvar(),
where this brings times down considerably.  For comparison,
`time man_lsfunc man*` took around 55 s (on my system) before this
change.  With groff(1), it would take around 14 s, and with mandoc(1)
(this patch), it takes 4 s.

Cc: Ingo Schwarze <schwarze@openbsd.org>
Cc: "G. Branden Robinson" <g.branden.robinson@gmail.com>
Cc: Gavin Smith <gavinsmith0123@gmail.com>
Cc: Dirk Gouders <dirk@gouders.net>
Cc: Colin Watson <cjwatson@debian.org>
Cc: Eli Zaretskii <eliz@gnu.org>
Cc: Larry McVoy <lm@mcvoy.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>