RELEASE, dist.mk: Use make(1)'s -B to force recreation of the dist files
Using FORCE unnecessarily restarts the entire build, even if we _know_
nothing changed.  That's boring.  Trust ourselves, and write the
commands in the RELEASE file as using '-B', to remind ourselves.
Forgetting to use -B will result in incorrect timestamps or versioning
in the distributed pages, so don't forget it ;).

While we're at it, let's also use -j4 directly, so I don't read the
paragraph reminding me to use -j _after_ I've already run it.  Let's
write -j4 instead of -j so that we don't crash some innocent's system.

Signed-off-by: Alejandro Colomar <alx@kernel.org>