dist.mk: dist: Don't pollute stderr unnecessarily
When running 'make dist' from a tarball (so we don't have git), we'll
see tons of errors saying we're not in a git tree.  However, that's not
meaningful, because that command is a no-op in such a scenario: the
(date) placeholder is not there anymore to be replaced.  Let's hide the
errors, unless V=1.

Signed-off-by: Alejandro Colomar <alx@kernel.org>