src.mk: Optimize pipeline
This pipeline is run for every make(1) invokation, and it was very slow,
due to running a new sed(1) process for every manual page.  Remove the
while loop, and rewrite so that all the commands in the pipeline are
only run once.  This brings times down ~10x in my system, from 1.5 s, to
just 0.14 s.

Signed-off-by: Alejandro Colomar <alx@kernel.org>