*/mk: mkdir -p (and install -d) are safe to use
There are no race conditions, contrary to what some sources say.
Previously, we had each directory depend on its parent directory, so
that only a few parent-most directories would be created with parents,
and all others would be created when their parents exist.  That was due
to FUD about TOCTOU race conditions with mkdir -p and parallel make, but
it was just FUD.

This patch makes it so that each directory is directly created with
parents without specifying any dependency.  This simplifies the makefile
at some cost: calling mkdir -p so many times is a bit slower.  However,
it's only about 5%, so we can live with it, since we're talking about
hundreds of a second for creating all the directories that this build
system knows about.

Reported-by: Ahelenia Ziemiańska <nabijaczleweli@nabijaczleweli.xyz>
Signed-off-by: Alejandro Colomar <alx@kernel.org>