*.mk: Use -wbreak in TROFFFLAGS, and -ww in NROFFFLAGS
We don't need to see all warnings everywhere we call troff(1).  Instead,
put all warnings in nroff mode, which we only run for the warnings, and
then only ask for warnings that depend on the output in other
invocations of troff(1).

-wbreak happens to enable only and all so-called "output warnings".  It
is an implementation detail of groff(1), but that's the best we can do
now.  If groff(1) changes their warnings organization, we'll need to
adapt to it.

Link: <https://lists.gnu.org/archive/html/groff/2023-05/msg00046.html>
Cc: "G. Branden Robinson" <g.branden.robinson@gmail.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>