gettimeofday.2: Add details about the nullability of tv in different libc's
tv arg is allowed to be NULL in bionic.

POSIX says this behavior is undefined, and bionic just exposes the Linux
syscall directly.  There's no code in bionic for
gettimeofday/settimeofday; just a description of the syscall name and
arguments from which an assembler stub is automatically generated at
build time.  musl and glibc go out of their way to behave differently
from the Linux kernel, but no idea why.

Signed-off-by: Zijun Zhao <zijunzhao@google.com>
Cc: Elliott Hughes <enh@google.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>