*.mk: Escape newlines instead of using +=
It's more robust to only spell the variable name once.  We've fixed typos
related to that in the past, and in this patch we're fixing another one:

	DEFAULT_CPPCHECKFLAgS += --inconclusive

Even if we could merge some of those lines into one (because they're
short), I prefer to keep one line per option.  Semantic newlines, aren't
they?  :)

Signed-off-by: Alejandro Colomar <alx@kernel.org>