string_copying.7: tfix
On some of the commas: There are a few of instances of

	Subject verb object partclause, advphrase.

For example:

	This function catenates the input character sequence
	| subject     | verb    | object                   |

	contained in a null-padded wixed-width buffer,
	| participial clause                        |

	into a destination string.
	| adverbial phrase       |

Imagining the participial clause away, there shouldn't be a comma
preceding the restrictive adverbial phrase: The input character sequence
is really, always catenated into a destination string; that is
essential. For example:

	This function catenates the input character sequence into
	a destination string.

The participial clause, being non-restrictive---there is but one input
character sequence that could be meant---, should be enclosed by commas.
That is the existing comma's purpose and doesn't work without the added,
first comma.

Cc: "G. Branden Robinson" <g.branden.robinson@gmail.com>
Signed-off-by: Lennart Jablonka <humm@ljabl.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>