*.mk: lint-c: Redirect error output to stderr
After this patch, you can do 2>/dev/null to only see one line per
Makefile recipe being run, which is nice and useful (for example, to
have an overview of which pages are failing, without having to care
about the specific errors).  I used this for example, to know which
files need to be touched to silence the known positives in `make lint`
in the Debian packaging is override_dh_test_auto.

In checkpatch(1), it's the program that writes to stdout.  Redirect it.

In clang-tidy(1) and cppcheck(1), it was a mistake of mine, since I
redirected to stdout to filter the output, and then forgot to redirect
back to stderr.

Fixes: acaa21bc4a6f ("Makefile, etc/clang-tidy/config.yaml: lint-clang-tidy: Add target to run clang-tidy(1) to lint example programs")
Fixes: eb1bdc8c117a ("Makefile: lint-iwyu: Be silent if there are no problems")
Signed-off-by: Alejandro Colomar <alx@kernel.org>