fsync.2: There are no writability requirements
Since Issue 3 (original release), fsync() was required to operate on
all valid fds. Since Issue 7 2018, fdatasync() is as well (and required
writability only by editorial mistake):
  https://www.austingroupbugs.net/view.php?id=501

"Some UNIXes require the fd to be writable" is a
needlessly-adversarial-to-the-user ‒
  https://101010.pl/@eater@cijber.social/110824211348995583
‒ way of saying "HP-UX and AIX have always been broken": just say that.

Originally appeared in 4.2BSD (4.1c.2BSD) so touch that as well since
we're mentioning the original interface.

Signed-off-by: Ahelenia Ziemiańska <nabijaczleweli@nabijaczleweli.xyz>
Cc: Jakub Wilk <jwilk@jwilk.net>
Cc: "G. Branden Robinson" <g.branden.robinson@gmail.com>
Cc: Guillem Jover <guillem@hadrons.org>
Cc: Sam James <sam@gentoo.org>
Signed-off-by: Alejandro Colomar <alx@kernel.org>