scripts/sortman: Add script to sort manual pages in book order
`LC_COLLATE=en_US.UTF-8` is needed to collate pages like _exit(3) next
to exit(3), instead of having all pages with a leading underscore
grouped at the beginning of a section.

intro(*), of course, is the fist page of each section, and subsections
go after the last page in the main corresponding section.

Signed-off-by: Alejandro Colomar <alx@kernel.org>