pthread_*.3: Remove AUTHOR section; add copyright; adapt TH
Xavier asked that his name be removed from the AUTHOR section, and since
we don't use that section in this project, we can just wipe it.

Add copyright notices to the pages, with Xavier as the original author,
and me, who will maintain them from now on.  Xavier ACKed the
Linux-man-pages-copyleft license.

Adapt the TH line:

-  Specify the date as (date).
-  Specify the project name in the 4th arg.
-  The first arg, the page name, must be the same as the file name.

Suggested-by: Xavier Leroy <Xavier.Leroy@inria.fr>
Signed-off-by: Alejandro Colomar <alx@kernel.org>