man*/: ffix (Use '.TQ' where appropriate)
When there are multiple tags for a paragraph, using a single TP and
separating the tags with commas makes the man(7) source more complex.
It also has a disadvantage: when searching through a manual page,
heuristics such as "   --option" don't work so well.

By using GNU's TQ, we simplify the source of the pages, and improve the
ability to search them.

Signed-off-by: Alejandro Colomar <alx@kernel.org>