scripts/LinuxManBook/prepare_linux_man_book.pl: Add new function BuildSec()
The intention is to move some code from BuildPage() into this function,
reducing the complexity of that inner function.  As a side effect, just
by adding this function, I've already noticed an optimization of 0.3 s.
The reason, I guess, is that it starts processing pages without waiting
for the entire sort to complete, which reduces the latency of the script.

Cc: Deri James <deri@chuzzlewit.myzen.co.uk>
Signed-off-by: Alejandro Colomar <alx@kernel.org>