scripts/LinuxManBook/: Simplify pipeline
Call the groff(1) pipeline only once.  This optimizes around 2 seconds,
while also simplifying the code.

This change was originally written by Deri, with some parts written by
Brian.  I took the script that Deri sent, and split it so that the
groff(1) pipeline is called from the shell script, and the Perl script
is limited to editing the man(7) pages.  This helps me understand the
process, since my understanding of Perl is very limited.  It also makes
this change smaller, so that it's less of a big-bang.

Link: <https://lore.kernel.org/linux-man/ZWkO4qPC4BxkwBNm@debian/T/#m3d453440b02dd189bc12d2e629c4026206025b40>
Co-developed-by: Deri James <deri@chuzzlewit.myzen.co.uk>
Co-developed-by: Brian Inglis <Brian.Inglis@Shaw.ca>
Signed-off-by: Alejandro Colomar <alx@kernel.org>