string_copying.7: Remove ustr2stp() and ustpcpy()
Users reported complains about having so many invented functions, and
the complexity of remembering all of them.

In these two cases, open-coding mempcpy(3) isn't so bad.  In fact, it's
quite readable.  It has the problem of type safety, since this function
accepts almost everything, but let users come up with a wrapper if they
need it; it's not like you would be calling these functions often.

Document how to use mempcpy(3) well in those cases.

Cc: "Serge E. Hallyn" <serge@hallyn.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>