ip.7: Some socket options are not supported by SOCK_STREAM
It was not made clear in several socket options that they were not
supported by SOCK_STREAM; this patch fixes that.

Socket options not supported by SOCK_STREAM are handled in the
ip_cmsg_recv_offset() function in <net/ipv4/ip_sockglue.c>.  The
function is called for udp sockets, and indirectly by ping and raw
sockets, but not for STREAM sockets, as they don't support these
options.

Link: <https://lore.kernel.org/linux-man/ejhphmjh74ebtk4br3id66f27a4yoh4aukrcz7m6dp7acsu6zr@crtueyadqzmp/T/#mb298ac7f71a348d1e6b423cfa32bfad9c28efa40>
Link: <https://lore.kernel.org/linux-man/ejhphmjh74ebtk4br3id66f27a4yoh4aukrcz7m6dp7acsu6zr@crtueyadqzmp/T/#u>
Signed-off-by: Oliver Crumrine <ozlinuxc@gmail.com>
Cc: Peter Seiderer <ps.report@gmx.net>
Signed-off-by: Alejandro Colomar <alx@kernel.org>