share/mk/: Remove support for installing compressed pages
I don't like compressed pages: they make it more difficult/slow to grep
them.  This was added to make it easier for users to install compressed
pages even when installing from source, but probably nobody cares at all
these days about a few KiB per page; drive space got cheap.  Also, some
of the algorithms were added just for benchmarking (for demonstrating
that they were bad in this case), and not for real use.

This reduces a use of compression tools, which seem to be dangerous
these days.

Link: <https://tukaani.org/xz-backdoor/>
Link: <https://cmpct.info/~sam/blog/posts/automatic-manpage-compression/>
Signed-off-by: Alejandro Colomar <alx@kernel.org>