proc.5: Refer to split-out manual pages for detailed description
Back in August 2023 various parts of proc(5) were split out into
separate manpages.  The final cleanup in 92cdcec79df0 ("proc.5: Clean up
after making sashimi of this page") missed to as well refer to the split
out manpages and retained the wording that details are found further
below in the manpages for the various files.

Fixes: 92cdcec79df0 ("proc.5: Clean up after making sashimi of this page")
Signed-off-by: Salvatore Bonaccorso <carnil@debian.org>
Signed-off-by: Alejandro Colomar <alx@kernel.org>